+++ ta4j-core/src/test/java/org/ta4j/core/indicators/AroonDownIndicatorTest.java
@@ -35,6 +35,7 @@ import org.junit.Test;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.BaseBarSeriesBuilder;
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.aroon.AroonDownIndicator;
 import org.ta4j.core.num.Num;
 
 public class AroonDownIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {